Password locking for more than three attempts, models package added to hold all classes
