+++ server/.gitignore
@@ -1,0 +1,3 @@
+/venv
+/app/app.log
+/app/log
