+++ server/app/requirements.txt
@@ -1,3 +1,4 @@
 fastapi==0.82.0
 uvicorn==0.18.3
 pydantic==1.10.2
+slowapi
